Se soluciona guardar en back

+++ nuxtendemo/nuxtendemo/src/main/java/com/demo/nuxtendemo/repository/UsuarioRepository.java
@@ -1,11 +1,10 @@
 package com.demo.nuxtendemo.repository;
-
 import com.demo.nuxtendemo.entitys.UsuariosEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<UsuariosEntity, Integer> {
+public interface UsuarioRepository extends JpaRepository<UsuariosEntity, Long> {
